# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Node.js Express Authentication and CRUD Operations",
      "version": "1.0.0",
      "description": "API documentation for Node.js Express Authentication and CRUD Operations"
    },
    "paths": {
      "/signup": {
        "post": {
          "summary": "Create a new user",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "User created successfully"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/login": {
        "post": {
          "summary": "Authenticate user and generate token",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Token generated successfully"
            },
            "401": {
              "description": "Invalid username or password"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/users": {
        "get": {
          "summary": "Get all users",
          "responses": {
            "200": {
              "description": "Success"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "post": {
          "summary": "Create a new user",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "User created successfully"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/users/{id}": {
        "put": {
          "summary": "Update a user",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User updated successfully"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "delete": {
          "summary": "Delete a user",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User deleted successfully"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      }
    }
  }
  